[VACGOM-120] test: 서비스 분리에 따른 application 계층 테스트 코드 분리
